your-ai-council/your-ai-council_server_python/requirements.txt
fastmcp>=0.0.13
fastapi>=0.115.0
uvicorn[standard]>=0.30.1
pydantic>=2.8.2
orjson>=3.10.7

